Add option to test and train Q-Agent
This update gives a choice to either add a Q-Agent or not. If a Q-Agent is added, we can either train them, or test them. Training of the Q-Agent stores no logs, and testing the Q-Agent doesn't affect the Q-Table of the agent. Also, when the "Add Bot" button is clicked, a Q-Agent is added if the Pickle file exists, and a naive agent is added if it doesn't.